expose display helpers
