+++ ts/packages/agentSdk/package.json
@@ -13,7 +13,9 @@
   "type": "module",
   "exports": {
     ".": "./dist/index.js",
-    "./helpers/commands": "./dist/commands.js"
+    "./helpers/action": "./dist/helpers/actionHelpers.js",
+    "./helpers/command": "./dist/helpers/commandHelpers.js",
+    "./helpers/display": "./dist/helpers/displayHelpers.js"
   },
   "types": "./dist/index.d.ts",
   "scripts": {
